Add render blog view
Modified the sidebar so when a user clicks on the blogs button it
performs a get request and renders the blogs on a separate view. Also
added a functional home button on the sidebar that will redirect the
user back to the home page. Next will be to incorporate comments into
the blogs.