Create show/hide blog comments function
Added show/hide blog comments function where when a user clicks on
the show comments button it renders the comments of that blog in the
proper div. The show comments button will then disappear and the hide
comments button will appear. When that button is clicked the comments
hide and the show comments button appears.