Edit blog template appends to main-body hidden
Made it so when the blogs are rendered the edit forms for each created
blogs render on the page but hidden. Next I will make it so when a user
clicks the edit button all of the rendered forms on the page disappear
and only the edit form for the blog you clicked on appears.