Fix blog create/update bug
So the bug was whenever a create/update action was successfully executed
the blog pages re-render but not with the changes created. To remedy
this issue I simply redirect the user to a intermediary page notifying
them of their successful creation/update with a button that performs a
get request to index all of the blogs and render them on the page.