Move event handlers from ui.js to events.js
This fixed a bug where the browser would load additional click handlers
to the browser each time the blogs were listed. The only bug remaining
is once I create a comment and open the create comment form again I get
a 422 Unprocessable Entity. If I re-render the blogs page again it works
the first time but after that the error persists until the blogs page
is re-rendered.