Add browser user token persistence
By storing the user hash given by the response from the browser in local
storage I am able to retrieve it upon page load and set it in the store
has in the store file. When a user logs out the key is deleted from
local storage. Next will be to refactor the code so that sign-in and
sign-out buttons appear properly if a token is present in local storage.