Add shopping cart button and modal
Add a static shopping cart modal on the top right of the page that (for
now) is at the top right corner no matter what. When a user clicks on
the button a modal pops up. Next, I will need to populate the modal
with the current contents of cart.items.