Order can be created on backend
Created functionality where an order can be posted to the backend.
Currently, the records parameter for orders is being posted as an array
of arrays where an example index would be "title: Rubber Soul". This is
not ideal as I would like it to be stored an an object with keys but
this is a good save point regardless.