Create Stripe charge functionality
Creating charges using the Stripe API is now fully functional. Next I
will need to style and add two more functions. One being deleting the
records that a user purchases after purchase. The second being once
a user add records to the cart and navigates to another section of the
site, when he/she returns to purchase more records and tries to add a
record to the cart that is already in the cart an error must display
and the record shouldn't be added to the cart. I am satisfied with my
progress with the Stripe API at this point so I am going to merge into
master and push.