Disable add record if record is already in cart
Added functionality that disables the ability to add a record to the
cart if the record is already in the cart. This was a problem because
if the user navigated away from the record purchase page and came back
they would be able to add a duplicate record to the cart. This added
functionality disables that.