Fix sign out failure bug
Fixed bug where if I were to nuke:pave while signed in and then try to
sign out it wouldn't let me due to storing the token in the store file
and the user token in local storage. Now if signing out fails the new
function signOutFailure will trigger and delete both items, allowing the
user to successfully sign out.