Fix clearCart bug
Fixed clear cart bug where only half of the cart items would clear after
a successful purchase was made. The bug was caused by a changing
cart.items.length as the for loop was running (it would reduce by one
each time a record was removed from the items array). I fixed it by
setting the cart.items.length to a variable before the for loop was run
so the number of times it ran through the array equaled the original
length of the cart.items array.